+++ src/com/oltpbenchmark/benchmarks/tpcc/pojo/District.java
@@ -18,6 +18,7 @@
 package com.oltpbenchmark.benchmarks.tpcc.pojo;
 
 import java.io.Serializable;
+import java.math.BigDecimal;
 
 public class District implements Serializable {
 
@@ -25,7 +26,7 @@ public class District implements Serializable {
 	public int d_w_id;
 	public int d_next_o_id;
 	public float d_ytd;
-	public float d_tax;
+	public BigDecimal d_tax;
 	public String d_name;
 	public String d_street_1;
 	public String d_street_2;
